requirements.txt
discord.py[voice]==0.16.6
PyDictionary==1.5.2
googletrans==2.2.0
aiohttp
websockets
chardet
youtube_dl
heroku3
